requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
cairocffi==1.6.1
CairoSVG==2.7.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
contourpy==1.1.1
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.43.1
frozenlist==1.4.0
greenlet==3.0.1
idna==3.4
kiwisolver==1.4.5
lxml==4.9.3
magic-filter==1.0.11
matplotlib==3.8.1
multidict==6.0.4
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
psycopg==3.1.14
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
reportlab==4.0.7
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
tinycss2==1.2.1
typing_extensions==4.7.1
urllib3==2.0.5
webencodings==0.5.1
yarl==1.9.2
